[refactor] BoorCampSummary 조회 pagination 추가
문제 사항 :
 - /api/v2/reviews/summary 엔드포인트에 pagination 추가 필요
변경 사항 :
 - pagination 추가 및 관련 로직 변경

+++ src/main/java/com/fasttime/domain/review/dto/response/BootcampReviewSummaryDTO.java
@@ -3,5 +3,5 @@ package com.fasttime.domain.review.dto.response;
 public record BootcampReviewSummaryDTO(
     String bootcamp,
     double averageRating,
-    int totalReviews
+    Long totalReviews
 ) {}
